\name{deCS-package}
\alias{deCS-package}
\alias{deCS}
\docType{package}
\title{Decoding the cell type specificity (deCS)
\packageTitle{deCS}
}
\description{A tool for scRNA-seq cell type annotation and elucidation of trait-cell associations for human complex traits.
\packageDescription{deCS}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{deCS}
\packageIndices{deCS}
}
\author{
\packageAuthor{deCS}

Maintainer: Guangsheng Pei
}
\references{
}
\keyword{ deCS; cell type }
\seealso{
https://github.com/GuangshengPei/deCS
}
\examples{
}
